Fix index-64bits.js (#254)
* Issue #245 - Add links to TP6 pages, merge 64bit and 32bit quantum pages.


* prep index-64bit.js to handle bot 32 and 64 bits

* first working version using grid

* remove index64 for plain index, match styles

* try to input variables

* fix parameter capture

* fixing props passing

* fix ?full option, add tp6 links (with preview)

* requested fixes

* fix brackets
