+++ src/quantum/config.js
@@ -1,67 +1,59 @@
 import zipObject from 'lodash/zipObject';
-import frum from '../utils/queryOps';
+import { frum } from '../utils/queryOps';
 
 
 const CONFIG = {
   windows64Regression: [
-    [
-      {
-        title: 'MotionMark HTML suite',
-        secondLink: '/quantum/windows10-64/motionmark_htmlsuite',
-        secondTitle: 'Breakdown',
-        signatures: { 'windows10-64': 'b7794dcfc9a7522e7c1615e4a59b13ee86c0d7a6' },
-      },
-      {
-        title: 'MotionMark Animometer',
-        secondLink: '/quantum/windows10-64/motionmark_animometer',
-        secondTitle: 'Breakdown',
-        signatures: { 'windows10-64': 'b34709c79b6de3d4e1a260a60da9717f20163d66' },
-      },
-    ],
-    [
-      {
-        title: 'ARES6',
-        secondLink: '/quantum/windows10-64/ARES6',
-        secondTitle: 'Breakdown',
-        signatures: { 'windows10-64': '22a0985bd873492d70d49c335a21fd49a73c87cf' },
-      },
-      {
-        title: 'JetStream',
-        secondLink: '/quantum/windows10-64/JetStream',
-        secondTitle: 'Breakdown',
-        signatures: { 'windows10-64': '799db0d801c7248924e93ca5a0b770020498da69' },
-      },
-    ],
+    {
+      title: 'MotionMark HTML suite',
+      secondLink: '/quantum/windows10-64/motionmark_htmlsuite',
+      secondTitle: 'Breakdown',
+      signatures: { 'windows10-64': 'b7794dcfc9a7522e7c1615e4a59b13ee86c0d7a6' },
+    },
+    {
+      title: 'MotionMark Animometer',
+      secondLink: '/quantum/windows10-64/motionmark_animometer',
+      secondTitle: 'Breakdown',
+      signatures: { 'windows10-64': 'b34709c79b6de3d4e1a260a60da9717f20163d66' },
+    },
+    {
+      title: 'ARES6',
+      secondLink: '/quantum/windows10-64/ARES6',
+      secondTitle: 'Breakdown',
+      signatures: { 'windows10-64': '22a0985bd873492d70d49c335a21fd49a73c87cf' },
+    },
+    {
+      title: 'JetStream',
+      secondLink: '/quantum/windows10-64/JetStream',
+      secondTitle: 'Breakdown',
+      signatures: { 'windows10-64': '799db0d801c7248924e93ca5a0b770020498da69' },
+    },
   ],
   windows32Regression: [
-    [
-      {
-        title: 'MotionMark HTML suite',
-        secondLink: '/quantum/windows7-32/motionmark_htmlsuite',
-        secondTitle: 'Breakdown',
-        signatures: { 'windows7-32': 'f90f8f88d6397a66fb9a586ddf19064f41ee3ce3' },
-      },
-      {
-        title: 'MotionMark Animometer',
-        secondLink: '/quantum/windows7-32/motionmark_animometer',
-        secondTitle: 'Breakdown',
-        signatures: { 'windows7-32': '4bbe842bb4e492b5f546291db455660287d10d84' },
-      },
-    ],
-    [
-      {
-        title: 'ARES6',
-        secondLink: '/quantum/windows7-32/ARES6',
-        secondTitle: 'Breakdown',
-        signatures: { 'windows7-32': '9585321f110f16414e272956c8892a38d03e7797' },
-      },
-      {
-        title: 'JetStream',
-        secondLink: '/quantum/windows7-32/JetStream',
-        secondTitle: 'Breakdown',
-        signatures: { 'windows7-32': '95535e0278558bd59ae324920ca0469278faadb4' },
-      },
-    ],
+    {
+      title: 'MotionMark HTML suite',
+      secondLink: '/quantum/windows7-32/motionmark_htmlsuite',
+      secondTitle: 'Breakdown',
+      signatures: { 'windows7-32': 'f90f8f88d6397a66fb9a586ddf19064f41ee3ce3' },
+    },
+    {
+      title: 'MotionMark Animometer',
+      secondLink: '/quantum/windows7-32/motionmark_animometer',
+      secondTitle: 'Breakdown',
+      signatures: { 'windows7-32': '4bbe842bb4e492b5f546291db455660287d10d84' },
+    },
+    {
+      title: 'ARES6',
+      secondLink: '/quantum/windows7-32/ARES6',
+      secondTitle: 'Breakdown',
+      signatures: { 'windows7-32': '9585321f110f16414e272956c8892a38d03e7797' },
+    },
+    {
+      title: 'JetStream',
+      secondLink: '/quantum/windows7-32/JetStream',
+      secondTitle: 'Breakdown',
+      signatures: { 'windows7-32': '95535e0278558bd59ae324920ca0469278faadb4' },
+    },
   ],
 
 };
