+++ src/utils/componentEnhancers.jsx
@@ -4,9 +4,9 @@ import GenericErrorBoundary from '../components/genericErrorBoundary';
 const wrapSectionComponentsWithErrorBoundaries = (sections = []) => (
   sections.map(section => ({
     ...section,
-    rows: section.rows.map(row => row.map((component, index) => (
+    rows: section.rows.map((component, index) => (
       <GenericErrorBoundary key={index}>{component}</GenericErrorBoundary>
-    ))),
+    )),
   }))
 );
